Change tag to NSInteger. Patch by Anibal Rindisbacher <anibal784@gmail.com>.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@30026 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSCell.h
@@ -304,8 +304,8 @@ enum {
 //
 // Assigning a Tag 
 //
-- (void)setTag:(int)anInt;
-- (int)tag;
+- (void)setTag:(NSInteger)anInt;
+- (NSInteger)tag;
 
 //
 // Formatting Data and Validating Input
